feat: group cli proc into folders
