+++ projects/cli/src/lib/cli/processors/system/cli-packages-command-processor.ts
@@ -12,8 +12,8 @@ import {
     initializeBrowserEnvironment,
     Package,
 } from '@qodalis/cli-core';
-import { ScriptLoaderService } from '../services/script-loader.service';
-import { CliPackageManagerService } from '../services/cli-package-manager.service';
+import { ScriptLoaderService } from '../../services/script-loader.service';
+import { CliPackageManagerService } from '../../services/cli-package-manager.service';
 
 @Injectable()
 export class CliPackagesCommandProcessor implements ICliCommandProcessor {
