+++ projects/cli/src/lib/cli/processors/system/index.ts
@@ -1,0 +1,9 @@
+export { CliPackagesCommandProcessor } from './cli-packages-command-processor';
+
+export { CliHotKeysCommandProcessor } from './cli-hot-keys-command-processor';
+
+export { CliHelpCommandProcessor } from './cli-help-command-processor';
+
+export { CliVersionCommandProcessor } from './cli-version-command-processor';
+
+export { CliHistoryCommandProcessor } from './cli-history-command-processor';
