+++ projects/cli/src/lib/cli/processors/users/index.ts
@@ -1,0 +1,2 @@
+export * from './cli-switch-user-command-processor';
+export * from './cli-whoami-command-processor';
